+++ packages/example/src/ports/element_repository_query_port.rs
@@ -1,0 +1,8 @@
+use async_trait::async_trait;
+use kti_cqrs_rs::errors::error::Error;
+
+#[async_trait]
+pub trait ElementRepositoryQueryPort: Send + Sync {
+  async fn get_elements(&self) -> Result<Vec<i32>, Error>;
+  async fn get_count(&self) -> Result<usize, Error>;
+}