+++ packages/example/src/ports/mod.rs
@@ -1,0 +1,3 @@
+pub mod element_repository_command_port;
+pub mod element_repository_query_port;
+pub mod element_service_port;