+++ packages/example/src/services/rw_lock/ports/mod.rs
@@ -1,2 +1,0 @@
-pub mod rw_lock_repository_port;
-pub mod rw_lock_service_port;