+++ packages/example/src/services/rw_lock/queries/mod.rs
@@ -1,2 +1,0 @@
-pub mod get_count_elements_query;
-pub mod get_elements_query;