+++ packages/kti_cqrs_rs/src/core/bus/mod.rs
@@ -1,3 +1,0 @@
-pub mod command_bus;
-pub mod query_bus;
-pub mod service_bus;