+++ packages/kti_cqrs_rs/src/lib.rs
@@ -1,2 +1,3 @@
-pub mod common;
-pub mod core;
+pub mod adapters;
+pub mod errors;
+pub mod ports;